transport: fix mem_estimate integer overflow
f.length is a uint32_t read off the wire, so f.length * 2 + 8000 wrapped
in 32 bits. A frame declaring a length near 2^32 produced a tiny estimate,
passed the max_request_size check, reserved almost no memory, and then had
its body read anyway. Only the first frame on a connection is size-checked
in read_frame(), so any later frame could do this.

Compute the estimate in size_t, and widen cql_server::_total_memory to
match (it truncated above 4 GiB anyway).

+++ transport/server.hh
@@ -238,7 +238,7 @@ private:
     netw::messaging_service& _ms;
     cql_server_config _config;
     semaphore& _memory_available;
-    const uint32_t _total_memory;
+    const size_t _total_memory;
     seastar::metrics::metric_groups _metrics;
     std::unique_ptr<event_notifier> _notifier;
 private:
